Fixed blending
If using blend function GL_SRC_ALPHA and GL_ONE_MINUS_SRC_ALPHA
we get color (RGBA) source factor (As,As,As,AS) and color dest
factor ((1,1,1) – (As,As,As), 1 – As). Well the problem is that
this results in resulting alpha for the final pixel < 1.0. Now
we use white background for the body which then shows up through
our GL renering. For example if src alpha is 0.6 then we will get
0.6 * 0.6 + 0.4 * 1.0 = 0.36 + 0.4 = 0.76 as the final pixel alpha
value.