Changes to allow dynamically changing source data.
Added names to sources.

Added a 'dynamicDraw' flag to the mapper.  If this is set, then the webGL buffers are created with gl.DYNAMIC_DRAW and can be changed.

Added an mapper.updateSourceBuffer method that can update the data of a named source.  If a Float32Array is passed in, no memory allocation is performed.