Copy buffers less and access buffers directly.
For the common use case of a source using a single array of data, take steps to avoid copying that data.  Prior to this change, when a source was set, a data buffer with a Float32Array was allocated for it.  This data buffer would then be inserted into the source, which internally would copy the Float32Array to a javascript array using a for loop.  Later, when the vertex buffer objects were created, the javascript array would be copied to a Float32Array.

Instead, if the source is not mutated, use the original Float32Array data buffer for the entire process.

This speeds up creation of data points in a test case by a factor of 3 (from an average time of 3.3 seconds down to 1.1 seconds for 300,000 points).

Additionally, the mapper object can return the internal Float32Array used for any source with a new getSourceBuffer method.  If this array is altered and dynamicDraw is turned on, then calling updateSourceBuffer without a data array will just reuse the source's internal Float32Array.  In this way, a calling program dos not need to create such an array itself.