Keep more precision in model view origins.
We have a special uniform called modelViewOriginUniform which offsets a model view matrix by an origin.  This is intended to reduce the GL computation error.  The origin was being stored in a vec3, which internally uses a GLMAT_ARRAY_TYPE, which, in turn, is a Float32Array.  This loses precision needlessly, as the origin doesn't get passed into WebGL but does get applied as a translation to the view matrix.  By using a native javascript array, the precision is maintained at whatever is initially supplied.