Clean up VBOs when we remove actors.
Eventually, we will run out of Vertex Buffer Objects.  Now, it is possible for the GPU to reuse them.  Some sources cite that it is poor practice to allocated and deallocate VBOs (though much better than never deallocating them), and it would be better to maintain a reference to them and reuse them instead.

Also, cache the depthBits parameter.  Getting WebGL parameters takes some time, and if we know they haven't changed, we can cache the values to get a minor speed improvement.