Perform pixel alignment before uniform origin shift.
On some layers, we use an origin shift to improve precision.  We often want pixel-perfect alignment at integer zoom levels.  Both of these features adjust the camera view matrix send to a shader.  If the origin is applied first, and a layer has a shifting origin (since we often will take the origin to be a vertex somewhere in that layer), then small variations in the origin computation can result in an exact 1 pixel jitter in the alignment rounding function.  By performing the alignment first, this does not happen.